Let's move this to `salt._compat`
