+++ salt/_compat.py
@@ -7,6 +7,7 @@ Salt compatibility code
 # Import python libs
 import sys
 import types
+import subprocess
 
 # Import 3rd-party libs
 from salt.ext.six import binary_type, string_types, text_type
@@ -122,3 +123,39 @@ def string_io(data=None):  # cStringIO can't handle unicode
         return cStringIO(bytes(data))
     except (UnicodeEncodeError, TypeError):
         return StringIO(data)
+
+
+if sys.version_info < (2, 7):
+
+    def check_output(*popenargs, **kwargs):  # pylint: disable=function-redefined
+        r"""Run command with arguments and return its output as a byte string.
+
+        If the exit code was non-zero it raises a CalledProcessError.  The
+        CalledProcessError object will have the return code in the returncode
+        attribute and output in the output attribute.
+
+        The arguments are the same as for the Popen constructor.  Example:
+
+        >>> check_output(["ls", "-l", "/dev/null"])
+        'crw-rw-rw- 1 root root 1, 3 Oct 18  2007 /dev/null\n'
+
+        The stdout argument is not allowed as it is used internally.
+        To capture standard error in the result, use stderr=STDOUT.
+
+        >>> check_output(["/bin/sh", "-c",
+        ...               "ls -l non_existent_file ; exit 0"],
+        ...              stderr=STDOUT)
+        'ls: non_existent_file: No such file or directory\n'
+        """
+        if 'stdout' in kwargs:
+            raise ValueError('stdout argument not allowed, it will be overridden.')
+        process = Popen(stdout=PIPE, *popenargs, **kwargs)
+        output, unused_err = process.communicate()
+        retcode = process.poll()
+        if retcode:
+            cmd = kwargs.get("args")
+            if cmd is None:
+                cmd = popenargs[0]
+            raise CalledProcessError(retcode, cmd, output=output)
+        return output
+    subprocess.check_output = check_output
